+++ src/main/java/com/example/dreampicturespring/entity/Ordertbl.java
@@ -17,12 +17,12 @@ public class Ordertbl {
     @Id //바로아래있는 멤버를 프라이머리키지정해준다.
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "Order_SEQ")//맴버테이블의 시퀀스를 만들어준다.
     @SequenceGenerator(sequenceName = "OrderSQ", allocationSize = 1, name = "Order_SEQ")//시퀀스를 만들어주는데 오른쪽의 시퀀스 이름을 똑같이 맞춰야한다.
-    private Long no_Order;
+    private Integer no_Order;
     private LocalDateTime Orderdate;
-    private Long orderstatus;
-    private Long no_painting;
-    private Long no_membership;
-    private Long totalprice;
+    private Integer orderstatus;
+    private Integer no_painting;
+    private Integer no_membership;
+    private Integer totalprice;
 
 
     public Ordertbl(){
