+++ src/main/java/com/example/dreampicturespring/repository/AdminRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface AdminRepository extends JpaRepository<Admintbl,Long> {//테이블이름,프라이머리키 데이터타입
+public interface AdminRepository extends JpaRepository<Admintbl,Integer> {//테이블이름,프라이머리키 데이터타입
 
     @Query(value = "select admintbl.adminpwd from admintbl where admintbl.adminid = :adminID ", nativeQuery = true)
     String findByadminID(@Param("adminID") String adminID);
