+++ src/main/java/com/example/dreampicturespring/repository/PaymentRepository.java
@@ -3,6 +3,6 @@ package com.example.dreampicturespring.repository;
 import com.example.dreampicturespring.entity.Paymenttbl;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentRepository extends JpaRepository<Paymenttbl,Long> {
+public interface PaymentRepository extends JpaRepository<Paymenttbl,Integer> {
 
 }
